+++ cli/gpgpdump/facade/facade_test.go
@@ -25,7 +25,7 @@ var (
 	resdataFromBindata1 = `Marker Packet (Obsolete Literal Packet) (tag 10) (3 bytes)
 	Literal data (3 bytes)
 Symmetric-Key Encrypted Session Key Packet (tag 3) (4 bytes)
-	Version: 4 (new)
+	Version: 4 (current)
 	Symmetric Algorithm: CAST5 (sym 3)
 	String-to-Key (S2K) Algorithm: Simple S2K (s2k 0)
 		Hash Algorithm: MD5 (hash 1)
@@ -33,7 +33,7 @@ Symmetrically Encrypted Data Packet (tag 9) (56 bytes)
 	Encrypted data (sym alg is specified in sym-key encrypted session key)
 `
 	resdataFromAscdata1 = `Signature Packet (tag 2) (94 bytes)
-	Version: 4 (new)
+	Version: 4 (current)
 	Signiture Type: Signature of a canonical text document (0x01)
 	Public-key Algorithm: ECDSA public key algorithm (pub 19)
 	Hash Algorithm: SHA256 (hash 8)
